# backend/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Grace ERP System — REST API
  description: |
    ## 🏫 Grace ERP System — Institutional Management Platform

    Complete REST API documentation for the **Grace ERP System** — a full-featured 
    Educational Resource Planning platform for schools and colleges.

    ### 🔐 Authentication
    All protected endpoints require a valid **JWT Bearer Token** in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ### ⚡ Rate Limiting
    Authentication endpoints are rate-limited to **5 requests per 15 minutes** per IP.

    ### 📌 Base URL
    - **Development**: `http://localhost:5000`
    - **Production**: *(configured via environment)*

  version: 1.0.0
  contact:
    name: Grace ERP Support
    email: support@grace-erp.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:5000/api
    description: Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
...
security:
  - BearerAuth: []
paths:
...
  /admission/direct-enroll:
    post:
      tags: [Admissions]
      summary: Direct enroll a new student
      description: |
        Directly enroll a student without going through the application process.
        Supports file uploads for admission documents.
        
        **Max file size**: 10 MB per file.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: Ananya Sharma
                class:
                  type: string
                  example: "9"
                section:
                  type: string
                  example: A
                academic_year_id:
                  type: integer
                  example: 3
                passportPhoto:
                  type: string
                  format: binary
                birthCert:
                  type: string
                  format: binary
                leavingCert:
                  type: string
                  format: binary
                casteCert:
                  type: string
                  format: binary
                aadharCopy:
                  type: string
                  format: binary
      responses:
        '201':
          description: Student enrolled
        '400':
          description: File too large or validation error
  /admission/all:
    get:
      tags: [Admissions]
      summary: Get all admission applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: academic_year_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of applications
...
  /admission/{id}:
    get:
      tags: [Admissions]
      summary: Get single application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application details
    put:
      tags: [Admissions]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, approved, rejected]
                remarks:
                  type: string
      responses:
        '200':
          description: Application updated
    delete:
      tags: [Admissions]
      summary: Delete application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application deleted

  # ════════════════════════════════════════════════════════════
  # ADMIN PROFILE
  # ════════════════════════════════════════════════════════════
...
  /timetable/class/{classroom_id}:
    get:
      tags: [Timetable]
      summary: Get timetable for a classroom
      parameters:
        - name: classroom_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Class timetable
  /timetable/teacher/{teacher_id}:
    get:
      tags: [Timetable]
      summary: Get timetable for a teacher
      parameters:
        - name: teacher_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Teacher timetable
  /timetable/slot:
    post:
      tags: [Timetable]
      summary: Save a timetable slot
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                classroom_id:
                  type: integer
                teacher_id:
                  type: integer
                subject_id:
                  type: integer
                day:
                  type: string
                  example: Monday
                period:
                  type: integer
                start_time:
                  type: string
                end_time:
                  type: string
      responses:
        '200':
          description: Slot saved
...
  /timetable/slot/{id}:
    delete:
      tags: [Timetable]
      summary: Delete a timetable slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Slot deleted

  # ════════════════════════════════════════════════════════════
  # TRANSPORT
  # ════════════════════════════════════════════════════════════
...